fix: KnowMorePage styles
