--- frontend/src/styles/ViolenceTypesPage.css
@@ -1,102 +1,0 @@
-.holepage {
-  overflow: auto; 
-  height: 80vh; 
-  text-align: left ;
-}
-  /* CSS sumário */
-  .summary{
-    padding: 5vh 5vh 0 5vh;
-  }
-  .summary-list{
-    list-style-type: none;
-    padding: 1vh 3vh;
-  }
-  .summary-button {
-    cursor: pointer;
-  }
-  .summary-button:hover{
-    color: purple;
-  }
-
-  /* CSS info  */
-  .infoMapadaViolencia{
-    padding: 5vh 5vh 0 5vh;
-    
-  }
-
-  .introduction{
-    padding: 2vh 5vh;
-    
-  }
-
-  .violenceMenu{
-    padding: 1vh 3vh;
-
-    .dropdownTitle {
-      font-weight: bold;
-      margin-top: -2vh;
-    }
-
-    .dropdownText {
-      margin-top: -2vh;
-      font-size: 15px;
-    }
-
-    .iconFisical{
-      color: #023AFF;
-      float: left;
-      margin-right: 2%;
-    }
-
-    .iconMoral{
-      color: #FA5A00;
-      float: left;
-      margin-right: 2%;
-    }
-
-    .iconPsychological{
-      color: #FCC605;
-      float: left;
-      margin-right: 2%;
-
-    }
-
-    .iconPatrimonial{
-      color: #248E12;
-      float: left;
-      margin-right: 2%;
-    }
-
-    .iconSexual{
-      color: #FD0000;
-      float: left;
-      margin-right: 2%;
-    }
-  }
-
-/* CSS seção FAQ */
-  .faq {
-    padding: 0 3vh 2vh 3vh;
-
-    .dropdownTitleFaq {
-      font-weight: bold;
-      margin-top: -2vh;
-    }
-
-    .dropdownTextFaq {
-      margin-top: -2vh;
-      font-size: 15px;
-    }
-    
-  }
-
-  .iconFaq {
-    float: left;
-    margin-right: 2%;
-    color: purple ;
-  }
-
-/* CSS Conheça Também */ 
-.LinksOtherOngs {
-  padding: 0vh 3vh;
-}
